moved packet proccessing into own type for more redability
