+++ src/net/udp.rs
@@ -2,8 +2,7 @@ use std::io::{self, Cursor, ErrorKind, Error, Read, Write};
 use std::net::{self, SocketAddr, ToSocketAddrs};
 
 use super::{SocketState, NetworkConfig};
-use packet::{header, Packet, FragmentBuffer, ReassemblyData};
-use self::header::{FragmentHeader, PacketHeader, HeaderReader};
+use packet::{header, Packet, PacketProcessor};
 use error::{NetworkError, Result};
 
 pub struct UdpSocket {
@@ -11,8 +10,7 @@ pub struct UdpSocket {
     state: SocketState,
     recv_buffer: [u8;1452],
     config: NetworkConfig,
-    /// buffer for temporarily fragment storage
-    reassembly_buffer: FragmentBuffer<ReassemblyData>
+    packet_processor: PacketProcessor
 }
 
 impl UdpSocket {
@@ -26,7 +24,7 @@ impl UdpSocket {
             socket,
             state,
             recv_buffer: [0;1452],
-            reassembly_buffer: FragmentBuffer::with_capacity(config.fragment_reassembly_buffer_size),
+            packet_processor: PacketProcessor::new(config.clone()),
             config: config,
         })
     }
@@ -36,22 +34,8 @@ impl UdpSocket {
 
         if len > 0 {
             let packet = self.recv_buffer[..len].to_owned();
-            let prefix_byte = packet[0];
-            let mut cursor = Cursor::new(packet);
 
-            let mut received_bytes = Ok(None);
-
-            if prefix_byte & 1 == 0 {
-                received_bytes = self.handle_normal_packet(&mut cursor, &_addr);
-            } else {
-                received_bytes = self.handle_fragment(&mut cursor);
-            }
-
-            return match received_bytes {
-                Ok(Some(payload)) => return Ok(Some(Packet::new(_addr, payload))),
-                Ok(None) => return Ok(None),
-                Err(e) => Err (Error::new(ErrorKind::Other, "Something whent wrong when when receiving"))
-            }
+            self.packet_processor.process_data(packet, _addr, &mut self.state)
         }else {
             return Err (Error::new(ErrorKind::Other, "Received bytes where of length 0"));
         }
@@ -78,96 +62,6 @@ impl UdpSocket {
 
     pub fn set_nonblocking(&mut self, nonblocking: bool) -> io::Result<()> {
         self.socket.set_nonblocking(nonblocking)
-    }
-
-    fn handle_fragment(&mut self, cursor: &mut Cursor<Vec<u8>>) -> Result<Option<Vec<u8>>>
-    {
-        // read fragment packet
-        let fragment_header = FragmentHeader::read(cursor)?;
-
-        // if fragment does not exists we need to insert a new entry
-        if !self.reassembly_buffer.exists(fragment_header.sequence) {
-            if fragment_header.id == 0 {
-                if fragment_header.packet_header.is_none() {
-                    return Err(NetworkError::AddConnectionToManagerFailed.into());
-                }
-
-                let packet_header = fragment_header.packet_header.unwrap();
-                let ack = packet_header.ack_seq;
-                let ack_bits = packet_header.ack_field;
-
-                let reassembly_data = ReassemblyData::new(fragment_header.sequence, ack, ack_bits, usize::from(fragment_header.num_fragments), fragment_header.size() as usize, 9 + self.config.fragment_size);
-
-                self.reassembly_buffer.insert(reassembly_data.clone(), fragment_header.sequence);
-            } else {
-                return Err(NetworkError::AddConnectionToManagerFailed.into());
-            }
-        }
-
-        let mut num_fragments_received = 0;
-        let mut num_fragments_total = 0;
-        let mut sequence = 0;
-        let mut total_buffer = Vec::new();
-
-        {
-            // get entry of previous received fragments
-            let reassembly_data = match self.reassembly_buffer.get_mut(fragment_header.sequence) {
-                Some(val) => val,
-                None => return Err(NetworkError::AddConnectionToManagerFailed.into())
-            };
-
-            // Got the data
-            if reassembly_data.num_fragments_total != usize::from(fragment_header.num_fragments) {
-                return Err(NetworkError::AddConnectionToManagerFailed.into());
-            }
-
-            if reassembly_data.fragments_received[usize::from(fragment_header.id)] {
-                return Err(NetworkError::AddConnectionToManagerFailed.into());
-            }
-
-            // increase number of received fragments and set the specific fragment to received.
-            reassembly_data.num_fragments_received += 1;
-            reassembly_data.fragments_received[usize::from(fragment_header.id)] = true;
-
-            // read payload after fragment header
-            let mut payload = Vec::new();
-            cursor.read_to_end(&mut payload)?;
-
-            // add the payload from the fragment to the buffer whe have in cache
-            reassembly_data.buffer.write(payload.as_slice());
-
-            num_fragments_received = reassembly_data.num_fragments_received;
-            num_fragments_total = reassembly_data.num_fragments_total;
-            sequence = reassembly_data.sequence as u16;
-            total_buffer = reassembly_data.buffer.clone();
-        }
-
-        // if whe received all fragments then remove entry and return the total received bytes.
-        if num_fragments_received == num_fragments_total {
-            let sequence = sequence as u16;
-            self.reassembly_buffer.remove(sequence);
-
-            return Ok(Some(total_buffer))
-        }
-
-        return Ok(None);
-    }
-
-    fn handle_normal_packet(&mut self, cursor: &mut Cursor<Vec<u8>>, addr: &SocketAddr) -> Result<Option<Vec<u8>>>
-    {
-        let packet_header = PacketHeader::read(cursor);
-
-        match packet_header {
-            Ok(header) => {
-                self.state.process_received(*addr, &header)?;
-
-                let mut payload = Vec::new();
-                cursor.read_to_end(&mut payload)?;
-
-                Ok(Some(payload))
-            }
-            Err(e) => {  return Err(NetworkError::HeaderParsingFailed.into());}
-        }
     }
 }
 
@@ -234,7 +128,7 @@ mod test {
         });
 
         let dummy_packet = Packet::new(addr, vec![123;4000]);
-        let send_result: io::Result<usize> = send_socket.send(dummy_packet);
+        let send_result = send_socket.send(dummy_packet);
         assert!(send_result.is_ok());
 
         handle.join();
