+++ src/packet/mod.rs
@@ -5,7 +5,9 @@ mod fragment_buffer;
 mod reassembly_data;
 mod packet_data;
 mod packet;
+mod packet_processor;
 
+pub use self::packet_processor::PacketProcessor;
 pub use self::reassembly_data::ReassemblyData;
 pub use self::packet_data::PacketData;
 pub use self::fragment_buffer::FragmentBuffer;
